# [Pretyping] Trigger CS search during pretyping of lambdas and applications