--- a/test-suite/success/CanonicalStructure.v
+++ b/test-suite/success/CanonicalStructure.v
@@ -33,7 +33,7 @@ Check (fun x : b => x != x).
 Inductive four := x0 | x1 | x2 | x3.
 Structure local := MKL { l : four }.
 
-Section X.
+Module X.
   Definition s0 := MKL x0.
   #[local] Canonical Structure s0.
   Check (refl_equal _ : l _ = x0).
@@ -48,11 +48,11 @@ End X.
 Fail Check (refl_equal _ : l _ = x0).
 Fail Check (refl_equal _ : l _ = x1).
 Fail Check (refl_equal _ : l _ = x2).
-Check s0.
-Check s1.
-Check s2.
+Check X.s0.
+Check X.s1.
+Check X.s2.
 
-Section Y.
+Module Y.
   Let s3 := MKL x3.
   Canonical Structure s3.
   Check (refl_equal _ : l _ = x3).
@@ -60,57 +60,59 @@ End Y.
 Fail Check (refl_equal _ : l _ = x3).
 Fail Check s3.
 
-Section V.
+Module V.
   #[canonical] Let s3 := MKL x3.
   Check (refl_equal _ : l _ = x3).
 End V.
 
-Section W.
+Module W.
   #[canonical, local] Definition s2' := MKL x2.
   Check (refl_equal _ : l _ = x2).
 End W.
 Fail Check (refl_equal _ : l _ = x2).
 
 (* Lambda keys *)
-Section L1.
+Module LambdaKeys.
   Structure cs_lambda := { cs_lambda_key : nat -> nat }.
-  #[local] Canonical Structure cs_lambda_func :=
-    {| cs_lambda_key := fun x => x + 1 |}.
-  Check (refl_equal _ : cs_lambda_key _ = fun _ => _ + _).
-End L1.
+  Module L1.
+    #[local] Canonical Structure cs_lambda_func :=
+      {| cs_lambda_key := fun x => x + 1 |}.
+    Check (refl_equal _ : cs_lambda_key _ = fun _ => _ + _).
+  End L1.
 
-Section L2.
-  #[local] Canonical Structure cs_lambda_func2 :=
-    {| cs_lambda_key := fun x => 1 + x |}.
-  Check (refl_equal _ : cs_lambda_key _ = fun x => 1 + x).
-End L2.
+  Module L2.
+    #[local] Canonical Structure cs_lambda_func2 :=
+      {| cs_lambda_key := fun x => 1 + x |}.
+    Check (refl_equal _ : cs_lambda_key _ = fun x => 1 + x).
+  End L2.
 
-Section L3.
-  #[local] Canonical Structure cs_lambda_func3 :=
-    {| cs_lambda_key := fun x => 1 + x |}.
-  Check (refl_equal _ : cs_lambda_key _ = Nat.add 1).
-End L3.
+  Module L3.
+    #[local] Canonical Structure cs_lambda_func3 :=
+      {| cs_lambda_key := fun x => 1 + x |}.
+    Check (refl_equal _ : cs_lambda_key _ = Nat.add 1).
+  End L3.
 
-Section L4.
-  #[local] Canonical Structure cs_lambda_func4 :=
-    {| cs_lambda_key := Nat.add 1 |}.
-  Check (refl_equal _ : cs_lambda_key _ = Nat.add 1).
-End L4.
+  Module L4.
+    #[local] Canonical Structure cs_lambda_func4 :=
+      {| cs_lambda_key := Nat.add 1 |}.
+    Check (refl_equal _ : cs_lambda_key _ = Nat.add 1).
+  End L4.
 
-Section L5.
-  #[local] Canonical Structure cs_lambda_func5 :=
-    {| cs_lambda_key := Nat.add 1 |}.
-  Check (refl_equal _ : cs_lambda_key _ = fun x => 1 + x).
-End L5.
+  Module L5.
+    #[local] Canonical Structure cs_lambda_func5 :=
+      {| cs_lambda_key := Nat.add 1 |}.
+    Check (refl_equal _ : cs_lambda_key _ = fun x => 1 + x).
+  End L5.
+End LambdaKeys.
 
-Section DepProd.
+Module DepProd.
   Structure hello := { hello_key : Type }.
-  Section FixedTypes.
+  Module FixedTypes.
     Local Canonical Structure hello_dep1 := {| hello_key := forall x : nat, x = x |}.
     Example ex_hello2 := let h := _ in fun f : hello_key h => (f : forall x : nat, x = x) 1.
   End FixedTypes.
 
-  Section VariableTypes.
+  Module VariableTypes.
     Local Canonical Structure hello_dep2 v1 v2 := {| hello_key := forall x : list v1, x = v2 |}.
     Example ex_hello1 : _ -> _ = nil :=
       let h := _ in fun f : hello_key h => (f : forall x : list _, _ = _) (@nil nat).
@@ -125,8 +127,8 @@ End DepProd.
    1. [f x : _] when [f : proj _]
    2. [(fun x => _) : proj _].
  *)
-Section NoCasts.
-  Section Basic.
+Module NoCasts.
+  Module Basic.
     Structure r1 (useless_param: bool) :=
       { #[canonical=no]  r1_pre : unit
       ; #[canonical=yes] r1_key : Type
@@ -137,7 +139,7 @@ Section NoCasts.
     Example ex_r1_2 p := let b := _ in (fun x => x) : @r1_key p b.
   End Basic.
 
-  Section Primitive.
+  Module Primitive.
     Local Set Primitive Projections.
     Structure r2 (useless_param: bool) :=
       { #[canonical=no]  r2_pre : unit
@@ -148,7 +150,7 @@ Section NoCasts.
     Example ex_r2_2 p := let b := _ in (fun x => x) : @r2_key p b.
   End Primitive.
 
-  Section UsedParameters.
+  Module UsedParameters.
     Structure r3 (useless_param: bool) (T : Type) :=
       { #[canonical=no]  r3_pre : unit
       ; #[canonical=yes] r3_key : Type
@@ -159,7 +161,7 @@ Section NoCasts.
     Example ex_r3_2 p := let b := _ in (fun x => x) : @r3_key p _ b.
   End UsedParameters.
 
-  Section LetBoundProjections.
+  Module LetBoundProjections.
     Structure r4 (useless_param: bool) :=
       { #[canonical=no]  r4_pre : unit
       ; #[canonical=no]  r4_let := true