<!-- Thank you for your contribution.
     Make sure you read the contributing guide and fill this template. -->

This PR allows the pretyper to look for suitable canonical instances to make the following typing problems succeed (which I consider to be dual to each other):
```
f : key ?c ⊢ f x : _
```
and

```
(fun _ : _ => _) : key ?c
```


I am actually calling unification to trigger the search because I couldn't figure out how to use the result of `lookup_canonical_projection`. There will surely be a cost to this. However, I did put in a check to see if there is a `Prod_cs` instance for `key`. If there isn't, unification is never called. This should limit the impact.

Unfortunately, there is also an unnecessary call to `whd_all` in `pretype_lambda`. This could be removed with some refactoring of `split_tycon`, where the same reduction is already performed. (One could also imagine that `split_tycon` would actually be a better place to trigger the CS search but that seemed wrong to me.)

I am marking this a draft because a) the code needs cleaning up, b) I am not convinced that unification is the best way to trigger CS here, and c) there might be metatheoretical implications to this change that I am not aware of.

**Depends on**: #12349 (because I didn't know how to create the evars for `_ -> _`, i.e. how to make sure the 2nd `_` doesn't depend on the first.)

<!-- Keep what applies -->
**Kind:** bug fix & feature.


<!-- If this is a bug fix, make sure the bug was reported beforehand. -->
Fixes / closes #11189


<!-- If there is a user-visible change in coqc/coqtop/coqchk/coq_makefile behavior and testing is not prohibitively expensive: -->
<!-- (Otherwise, remove this line.) -->
- [X] Added / updated test-suite
<!-- If this is a feature pull request / breaks compatibility: -->
<!-- (Otherwise, remove these lines.) -->
- [ ] Corresponding documentation was added / updated (including any warning and error messages added / removed / modified).
- [ ] Entry added in the changelog (see https://github.com/coq/coq/tree/master/doc/changelog#unreleased-changelog for details).