@Janno 2020-05-22
Since the benchmarking server seems rather underused right now I decided to start a benchmark to find out how bad this naive approach really is in terms of performance impact:
https://ci.inria.fr/coq/job/benchmark-part-of-the-branch/944/