@JasonGross 2020-05-28
I believe there's a real slowdown here, of approximately the same size you saw before.  I've just added a lot more files to engine-bench in the meantime, so the absolute times are 4x as big, and the % diff is 4x as small (this is an issue with monolithic performance experiments).  If you look at the absolute time difference, previously it was 1148.55 - 1119.66 = 28.89, and now it is 4345.44 - 4318.32 = 27.12, basically the same.  The [app_n file here](https://ci.inria.fr/coq/job/benchmark-part-of-the-branch/ws/948/html/coq-engine-bench/coq/PerformanceDemos/app_n.v.html) also shows the same slowdown on `Axiom` as it did before.  By the way, if you look at the internal timing logs of [the new version](https://ci.inria.fr/coq/job/benchmark-part-of-the-branch/ws/948/opam.NEW/ocaml-base-compiler.4.07.1/.opam-switch/build/coq-engine-bench.dev/coq/PerformanceExperiments/app_n_ltac2/_0_Fast.log) vs [the old version](https://ci.inria.fr/coq/job/benchmark-part-of-the-branch/ws/948/opam.OLD/ocaml-base-compiler.4.07.1/.opam-switch/build/coq-engine-bench.dev/coq/PerformanceExperiments/app_n_ltac2/_0_Fast.log), you can see a small but consistent slowdown of about 5%:
NEW:
Params: n= 20000
Tactic call build ran for 0. secs (0.u,0.s) (success)
Tactic call typecheck ran for 2.756 secs (2.756u,0.s) (success)
Tactic call build-and-typecheck ran for 2.757 secs (2.756u,0.s) (success)
Tactic call refine ran for 2.763 secs (2.763u,0.s) (success)

OLD:
Params: n= 20000
Tactic call build ran for 0. secs (0.u,0.s) (success)
Tactic call typecheck ran for 2.626 secs (2.627u,0.s) (success)
Tactic call build-and-typecheck ran for 2.627 secs (2.627u,0.s) (success)
Tactic call refine ran for 2.63 secs (2.627u,0.s) (success)

If you want to see this behavior locally with engine-bench, you can clone it, and then run
```coq
cd engine-bench/coq/PerformanceExperiments
make app_n_ltac2/_0_Fast.log
```
or if you want to play with it interactively, you can look at https://github.com/mit-plv/engine-bench/blob/master/coq/PerformanceExperiments/app_n_ltac2.v and issue `run0 Fast.`, or else something like
```coq
Goal True.
  let n := constr:(12700) in
  let g := constr:(goal n) in
  assert g; [ red_goal n; time_solve_goal0 n | ].
```
Note that the Ltac2 version I point you at here has much cleaner and more reliable timing than the uconstr / Ltac version:

uconstr | Ltac2
--|--
<img src="https://mit-plv.github.io/engine-bench/coq/app-n-uconstr.svg" height=100px /> | <img src="https://mit-plv.github.io/engine-bench/coq/app-n-ltac2.svg" height=100px />